SmartPlantMonitoringSystem/gcp/requirements.txt
tensorflow==2.17.0
keras==3.5.0
google-cloud-storage==2.16.0 
Pillow==10.4.0 
matplotlib==3.9.2
numpy==1.26.4